Update and refactor recalc-sessions scripts
