+++ scripts/__KT_DATABASE_MIGRATION/state-sync/sync-state.ts
@@ -1,7 +1,7 @@
 import CreateLogCtx from "lib/logger/logger";
 import { RecalcAllScores, UpdateAllPBs } from "utils/calculations/recalc-scores";
+import { RecalcSessions } from "utils/calculations/recalc-sessions";
 import { CreateGameProfiles } from "./create-game-profiles";
-import { RecalcSessions } from "./recalc-sessions";
 
 const logger = CreateLogCtx(__filename);
